Challenge7/app/requirements.txt
openai==1.30.1
azure-search-documents==11.6.0b1
pillow==10.3.0
pandas==2.2.2
tiktoken==0.7.0
chainlit==1.1.201
